[datastore] Move flags and ConnectParameters to a dedicated package, merge datastoreutils into datastore
