+++ cmds/db-manager/migration/migrate.go
@@ -12,7 +12,7 @@ import (
 
 	"github.com/coreos/go-semver/semver"
 	"github.com/interuss/dss/pkg/datastore"
-	datastoreflags "github.com/interuss/dss/pkg/datastore/flags"
+	"github.com/interuss/dss/pkg/datastore/params"
 
 	"github.com/interuss/stacktrace"
 	"github.com/spf13/cobra"
@@ -227,7 +227,7 @@ func migrate(cmd *cobra.Command, _ []string) error {
 
 func connectTo(ctx context.Context, dbName string) (*datastore.Datastore, error) {
 	// Connect to database server
-	connectParameters := datastoreflags.ConnectParameters()
+	connectParameters := params.GetConnectParameters()
 	connectParameters.DBName = dbName
 	return datastore.Dial(ctx, connectParameters)
 }
