+++ pkg/datastore/datastore.go
@@ -3,12 +3,17 @@ package datastore
 import (
 	"context"
 	"fmt"
+	"strings"
 	"time"
 
 	"github.com/coreos/go-semver/semver"
 	"github.com/exaring/otelpgx"
+	"github.com/interuss/dss/pkg/datastore/params"
+	"github.com/interuss/dss/pkg/logging"
 	"github.com/interuss/stacktrace"
 	"github.com/jackc/pgx/v5/pgxpool"
+	"github.com/robfig/cron/v3"
+	"go.uber.org/zap"
 )
 
 type Datastore struct {
@@ -16,9 +21,64 @@ type Datastore struct {
 	Pool    *pgxpool.Pool
 }
 
+const (
+	CodeRetryable = stacktrace.ErrorCode(1)
+)
+
 var UnknownVersion = &semver.Version{}
 
-func Dial(ctx context.Context, connParams ConnectParameters) (*Datastore, error) {
+func checkDatabase(ctx context.Context, db *Datastore, databaseName string) {
+	logger := logging.WithValuesFromContext(ctx, logging.Logger)
+	statsPtr := db.Pool.Stat()
+	if int(statsPtr.TotalConns()) == 0 {
+		logger.Warn("Failed periodic DB Ping (TotalConns=0)", zap.String("Database", databaseName))
+	} else {
+		logger.Info("Successful periodic DB Ping", zap.String("Database", databaseName))
+	}
+}
+
+func DialStore[S any](ctx context.Context, dbName string, withCheckCron bool, newStore func(*Datastore) (S, error)) (S, error) {
+
+	var zero S
+
+	cp := params.GetConnectParameters()
+	cp.DBName = dbName
+
+	db, err := Dial(ctx, cp)
+
+	if err != nil {
+		if strings.Contains(err.Error(), "connect: connection refused") {
+			return zero, stacktrace.PropagateWithCode(err, CodeRetryable, "Failed to connect to datastore server for %s", dbName)
+		}
+		return zero, stacktrace.Propagate(err, "Failed to connect to %s database", dbName)
+	}
+	s, err := newStore(db)
+	if err != nil {
+		db.Pool.Close()
+		if strings.Contains(err.Error(), "connect: connection refused") || strings.Contains(err.Error(), fmt.Sprintf("database \"%s\" does not exist", dbName)) || strings.Contains(err.Error(), "database has not been bootstrapped with Schema Manager") {
+			return zero, stacktrace.PropagateWithCode(err, CodeRetryable, "Failed to create %s store", dbName)
+		}
+		return zero, stacktrace.Propagate(err, "Failed to create %s store", dbName)
+	}
+
+	if withCheckCron {
+		c := cron.New()
+		if _, err := c.AddFunc("@every 1m", func() { checkDatabase(ctx, db, dbName) }); err != nil {
+			db.Pool.Close()
+			return zero, stacktrace.Propagate(err, "Failed to schedule db check for %s", dbName)
+		}
+		c.Start()
+
+		go func() {
+			<-ctx.Done()
+			c.Stop()
+		}()
+	}
+
+	return s, nil
+}
+
+func Dial(ctx context.Context, connParams params.ConnectParameters) (*Datastore, error) {
 	dsn, err := connParams.BuildDSN()
 	if err != nil {
 		return nil, stacktrace.Propagate(err, "Failed to create connection config for pgx")
