--- pkg/datastore/flags/flags.go
@@ -1,38 +1,0 @@
-package flags
-
-import (
-	"flag"
-
-	"github.com/interuss/dss/pkg/datastore"
-)
-
-var (
-	connectParameters datastore.ConnectParameters
-)
-
-// ConnectParameters returns a ConnectParameters instance that gets populated from well-known CLI flags.
-func ConnectParameters() datastore.ConnectParameters {
-	return connectParameters
-}
-
-func init() {
-	flag.StringVar(&connectParameters.ApplicationName, "datastore_application_name", "dss", "application name for tagging the connection to the database")
-	flag.StringVar(&connectParameters.Host, "datastore_host", "", "database host to connect to")
-	flag.IntVar(&connectParameters.Port, "datastore_port", 26257, "database port to connect to")
-	flag.StringVar(&connectParameters.SSL.Mode, "datastore_ssl_mode", "disable", "database sslmode")
-	flag.StringVar(&connectParameters.SSL.Dir, "datastore_ssl_dir", "", "directory to ssl certificates. Must contain files: ca.crt, client.<user>.crt, client.<user>.key")
-	flag.StringVar(&connectParameters.Credentials.Username, "datastore_user", "root", "database user to authenticate as")
-	flag.IntVar(&connectParameters.MaxOpenConns, "datastore_max_open_conns", 4, "maximum number of open connections to the database, default is 4")
-	flag.IntVar(&connectParameters.MaxConnIdleSeconds, "datastore_max_conn_idle_secs", 30, "maximum amount of time in seconds a connection may be idle, default is 30 seconds")
-	flag.IntVar(&connectParameters.MaxRetries, "datastore_max_retries", 100, "maximum number of attempts to retry a query in case of contention, default is 100")
-
-	flag.StringVar(&connectParameters.ApplicationName, "cockroach_application_name", connectParameters.ApplicationName, "DEPRECATED: use 'datastore_application_name' instead")
-	flag.StringVar(&connectParameters.Host, "cockroach_host", connectParameters.Host, "DEPRECATED: use 'datastore_host' instead")
-	flag.IntVar(&connectParameters.Port, "cockroach_port", connectParameters.Port, "DEPRECATED: use 'datastore_port' instead")
-	flag.StringVar(&connectParameters.SSL.Mode, "cockroach_ssl_mode", connectParameters.SSL.Mode, "DEPRECATED: use 'datastore_ssl_mode' instead")
-	flag.StringVar(&connectParameters.SSL.Dir, "cockroach_ssl_dir", connectParameters.SSL.Dir, "DEPRECATED: use 'datastore_ssl_dir' instead")
-	flag.StringVar(&connectParameters.Credentials.Username, "cockroach_user", connectParameters.Credentials.Username, "DEPRECATED: use 'datastore_user' instead")
-	flag.IntVar(&connectParameters.MaxOpenConns, "max_open_conns", connectParameters.MaxOpenConns, "DEPRECATED: use 'datastore_max_open_conns' instead")
-	flag.IntVar(&connectParameters.MaxConnIdleSeconds, "max_conn_idle_secs", connectParameters.MaxConnIdleSeconds, "DEPRECATED: use 'datastore_max_conn_idle_secs' instead")
-	flag.IntVar(&connectParameters.MaxRetries, "cockroach_max_retries", connectParameters.MaxRetries, "DEPRECATED: use 'datastore_max_retries' instead")
-}
